+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file
 ###Added
 - Insecure option to the REST call constructs
 - (Un)escape XML constructs
+- More descriptive information in the robot status bar
+
 ###Fixed
 - Issue where focus would be lost on the project view [CTC-1254](https://xillio.atlassian.net/browse/CTC-1254)